Code/Face Recognization/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
opencv-python==4.12.0.88
ultralytics==8.3.5
torch==2.4.1
onnxruntime==1.23.2
numpy==2.2.6
python-multipart==0.0.9
paho-mqtt==1.6.1